+++ app/Http/Requests/NewAdminRequest.php
@@ -6,7 +6,7 @@ use App\Models\Admin;
 use App\Models\User;
 use Illuminate\Foundation\Http\FormRequest;
 
-class EditAdminValidateRequest extends FormRequest
+class NewAdminRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -28,16 +28,14 @@ class EditAdminValidateRequest extends FormRequest
         return [
             "login_id" => [
                 "required",
-                function ($value, $attribute , $fail) {
-                    if ($value != User::where('id', $this->user)->first()) { //if change login_id
-                        if (User::where('login_id', $value)->count() > 0 || Admin::where('login_id', $value)
-                                                                                 ->count() > 0) {
-                            $fail('Login ID has exists!');
-                        }
+                function ($attribute, $value, $fail) {
+                    if (User::where('login_id', $value)->count() > 0 || Admin::where('login_id', $value)->count() > 0) {
+                        $fail('Login ID has exists!');
                     }
                 },
             ],
-            "email"    => "required|unique:admins,email,".$this->admin,
+            "email"    => "required|unique:admins,email",
+            "password" => "required",
         ];
     }
 
@@ -47,6 +45,7 @@ class EditAdminValidateRequest extends FormRequest
             "login_id.required" => __("validation.required"),
             "email.required"    => __("validation.required"),
             "email.unique"      => __("validation.unique"),
+            "password.required" => __("validation.required"),
         ];
     }
 }
