+++ app/Http/Requests/NewUserRequest.php
@@ -6,7 +6,7 @@ use App\Models\Admin;
 use App\Models\User;
 use Illuminate\Foundation\Http\FormRequest;
 
-class EditUserValidateRequest extends FormRequest
+class NewUserRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -28,18 +28,15 @@ class EditUserValidateRequest extends FormRequest
         return [
             "login_id" => [
                 "required",
-                function ($value, $attribute , $fail) {
-                    if ($value != User::where('id', $this->user)->first()) { //if change login_id
-                        if (User::where('login_id', $value)->count() > 0 || Admin::where('login_id', $value)
-                                                                                 ->count() > 0) {
-                            $fail('Login ID has exists!');
-                        }
+                function ($attribute, $value, $fail) {
+                    if (User::where('login_id', $value)->count() > 0 || Admin::where('login_id', $value)->count() > 0) {
+                        $fail('Login ID has exists!');
                     }
                 },
             ],
-            "email"    => "required|unique:users,email,".$this->user,
+            "email"    => "required|unique:users,email",
+            "password" => "required",
             "start_at" => "required",
-            "end_at"   => "nullable|after:start_at",
             "code"     => "required",
         ];
     }
@@ -52,7 +49,6 @@ class EditUserValidateRequest extends FormRequest
             "email.unique"      => __("validation.unique"),
             "password.required" => __("validation.required"),
             "start_at.required" => __("validation.required"),
-            "end_at.after"      => __("validation.after"),
             "code.required"     => __("validation.required"),
         ];
     }
